requirements.txt
# Core
python-docx
pydantic
uuid

# MCP Framework
strands
mcp

# Google Vertex AI
google-cloud-aiplatform
vertexai

# Optional Utilities
requests


fastapi
uvicorn
python-dotenv
